+++ services/api-server/src/simcore_service_api_server/models/domain/users.py
@@ -26,5 +26,7 @@ class UserInDB(BaseModel):
     status: UserStatus
     role: UserRole
 
+    # TODO: connect name <-> first_name, last_name
+
     class Config:
         orm_mode = True
